+++ addons/yat/src/commands/Man.cs
@@ -10,13 +10,15 @@ namespace YAT.Commands;
 
 [Command("man", "Displays the manual for a command.", "[b]Usage[/b]: man [i]command_name[/i]")]
 [Argument("command_name", "string", "The name of the command to display the manual for.")]
+[Option("-e", "bool", "Embed the manual in the terminal. Instead of opening in a new window.")]
 public sealed class Man : ICommand
 {
 	private readonly LRUCache<string, string> cache = new(10);
 
 	public CommandResult Execute(CommandData data)
 	{
-		var commandName = data.RawData[1];
+		var commandName = (string)data.Arguments["command_name"];
+		var embed = (bool)data.Options["-e"];
 
 		if (!RegisteredCommands.Registered.TryGetValue(commandName, out Type value))
 			return ICommand.InvalidCommand(Messages.UnknownCommand(commandName));
@@ -40,7 +42,8 @@ public sealed class Man : ICommand
 
 		cache.Add(commandName, manual);
 
-		data.Terminal.Print(manual);
+		if (embed) data.Terminal.Print(manual);
+		else data.Terminal.FullWindowDisplay.Open(manual);
 
 		return ICommand.Success();
 	}